# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Metis API",
    "description": "Metis API documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/v1/api/users/channels": {
      "get": {
        "tags": ["Channels"],
        "summary": "Get all user channels",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetChannelsResponse"
            }
          },
          "403": {
            "description": "Not authenticated",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChannelRecord": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "passphrase": {
          "type": "string"
        },
        "account": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "publicKey": {
          "type": "string"
        },
        "sender": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "date": {
          "type": "number"
        },
        "confirmed": {
          "type": "boolean"
        }
      }
    },
...
    "Channel": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "channel_record": {
          "type": "object",
          "$ref": "#/definitions/ChannelRecord"
        },
        "date": {
          "type": "number"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "GetChannelsResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "channels": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Channel"
          }
        },
        "total_channels_number": {
          "type": "number"
        }
      }
    },
...
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message":{
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}